more CWL objects added
