+++ lib/core/cwl/requirements.go
@@ -3,75 +3,39 @@ package cwl
 import (
 	"errors"
 	"fmt"
+	requirements "github.com/MG-RAST/AWE/lib/core/cwl/requirements"
 	"github.com/MG-RAST/AWE/lib/logger"
-	"github.com/davecgh/go-spew/spew"
-	"github.com/mitchellh/mapstructure"
+	//"github.com/davecgh/go-spew/spew"
+	//"github.com/mitchellh/mapstructure"
 )
 
 type Requirement interface {
 	GetClass() string
 }
 
-type StepInputExpressionRequirement struct {
-	//Class         string `yaml:"class"`
-}
-
-func (c StepInputExpressionRequirement) GetClass() string { return "StepInputExpressionRequirement" }
-func (c StepInputExpressionRequirement) GetId() string    { return "None" }
-
-type DockerRequirement struct {
-	//Class         string `yaml:"class"`
-	DockerPull    string `yaml:"dockerPull"`
-	DockerLoad    string `yaml:"dockerLoad"`
-	DockerFile    string `yaml:"dockerFile"`
-	DockerImport  string `yaml:"dockerImport"`
-	DockerImageId string `yaml:"dockerImageId"`
-}
-
-func (c DockerRequirement) GetClass() string { return "DockerRequirement" }
-func (c DockerRequirement) GetId() string    { return "None" }
-
-type ShockRequirement struct {
-	Host string `yaml:"host"`
-}
-
-func (s ShockRequirement) GetClass() string { return "ShockRequirement" }
-func (s ShockRequirement) GetId() string    { return "None" }
-
 func NewRequirement(class string, obj interface{}) (r Requirement, err error) {
 	switch class {
 	case "DockerRequirement":
-		var requirement DockerRequirement
-		err = mapstructure.Decode(obj, &requirement)
-		if err != nil {
-			spew.Dump(obj)
-			err = fmt.Errorf("object not a DockerRequirement")
-			return
-
-		}
-		r = requirement
+		r, err = requirements.NewDockerRequirement(obj)
+		return
 	case "StepInputExpressionRequirement":
-		var requirement StepInputExpressionRequirement
-		r = requirement
+		r, err = requirements.NewStepInputExpressionRequirement(obj)
+		return
 	case "ShockRequirement":
-		var requirement ShockRequirement
-		err = mapstructure.Decode(obj, &requirement)
-		if err != nil {
-			spew.Dump(obj)
-			err = fmt.Errorf("object not a DockerRequirement")
-			return
-
-		}
-		r = requirement
+		r, err = requirements.NewShockRequirement(obj)
+		return
+	case "InitialWorkDirRequirement":
+		r, err = requirements.NewInitialWorkDirRequirement(obj)
+		return
 	default:
-		err = errors.New("object class not supported " + class)
+		err = errors.New("Requirement class not supported " + class)
 
 	}
 	return
 }
 
 // create Requirement object from interface type
-func CreateRequirement(class string, v interface{}) (requirement Requirement, err error) {
+func CreateRequirementDEPRECATED(class string, v interface{}) (requirement Requirement, err error) {
 	switch v.(type) {
 	case map[interface{}]interface{}: // the Requirement is a struct itself
 		logger.Debug(1, "(CreateRequirementArray) type is map[interface{}]interface{}")
@@ -110,7 +74,7 @@ func CreateRequirementArray(original interface{}) (new_array_ptr *[]Requirement,
 			//var requirement Requirement
 			class := k.(string)
 
-			requirement, xerr := CreateRequirement(class, v)
+			requirement, xerr := NewRequirement(class, v)
 			if xerr != nil {
 				err = xerr
 				return
@@ -128,7 +92,7 @@ func CreateRequirementArray(original interface{}) (new_array_ptr *[]Requirement,
 			}
 			class := empty.GetClass()
 
-			requirement, xerr := CreateRequirement(class, v)
+			requirement, xerr := NewRequirement(class, v)
 			if xerr != nil {
 				err = xerr
 				return
