+++ lib/core/cwl/workflow.go
@@ -5,7 +5,7 @@ import (
 	"github.com/davecgh/go-spew/spew"
 	"github.com/mitchellh/mapstructure"
 	//"os"
-	"reflect"
+	//"reflect"
 	//"strings"
 )
 
@@ -28,176 +28,6 @@ func (w *Workflow) SetId(id string)  { w.Id = id }
 func (w *Workflow) is_CWL_minimal()  {}
 func (w *Workflow) is_Any()          {}
 
-type WorkflowStep struct {
-	Id            string               `yaml:"id"`
-	In            []WorkflowStepInput  `yaml:"in"` // array<WorkflowStepInput> | map<WorkflowStepInput.id, WorkflowStepInput.source> | map<WorkflowStepInput.id, WorkflowStepInput>
-	Out           []WorkflowStepOutput `yaml:"out"`
-	Run           string               `yaml:"run"` // Specification unclear: string | CommandLineTool | ExpressionTool | Workflow
-	Requirements  []Requirement        `yaml:"requirements"`
-	Hints         []Requirement        `yaml:"hints"`
-	Label         string               `yaml:"label"`
-	Doc           string               `yaml:"doc"`
-	Scatter       string               `yaml:"scatter"`       // ScatterFeatureRequirement
-	ScatterMethod string               `yaml:"scatterMethod"` // ScatterFeatureRequirement
-}
-
-func (w WorkflowStep) GetOutput(id string) (output *WorkflowStepOutput, err error) {
-	for _, o := range w.Out {
-		// o is a WorkflowStepOutput
-		if o.Id == id {
-			output = &o
-			return
-		}
-	}
-	err = fmt.Errorf("WorkflowStepOutput %s not found in WorkflowStep", id)
-	return
-}
-
-//http://www.commonwl.org/v1.0/Workflow.html#WorkflowStepInput
-type WorkflowStepInput struct {
-	Id        string          `yaml:"id"`
-	Source    []string        `yaml:"source"` // MultipleInputFeatureRequirement
-	LinkMerge LinkMergeMethod `yaml:"linkMerge"`
-	Default   Any             `yaml:"default"`   // type Any does not make sense
-	ValueFrom Expression      `yaml:"valueFrom"` // StepInputExpressionRequirement
-}
-
-func (w WorkflowStepInput) GetClass() string { return "WorkflowStepInput" }
-func (w WorkflowStepInput) GetId() string    { return w.Id }
-func (w WorkflowStepInput) SetId(id string)  { w.Id = id }
-
-func (w WorkflowStepInput) is_CWL_minimal() {}
-
-//func (input WorkflowStepInput) GetString() (value string, err error) {
-//	if len(input.Source) > 0 {
-//		err = fmt.Errorf("Source is defined and should be used")
-//	} else if string(input.ValueFrom) != "" {
-//		value = string(input.ValueFrom)
-//	} else if input.Default != nil {
-//		value = input.Default
-//	} else {
-//		err = fmt.Errorf("no input (source, default or valueFrom) defined for %s", id)
-//	}
-//	return
-//}
-
-func (input WorkflowStepInput) GetObject(c *CWL_collection) (obj *CWL_object, err error) {
-
-	var cwl_obj CWL_object
-
-	if len(input.Source) > 0 {
-		err = fmt.Errorf("Source is defined and should be used")
-	} else if string(input.ValueFrom) != "" {
-		new_string := string(input.ValueFrom)
-		evaluated_string := c.Evaluate(new_string)
-
-		cwl_obj = &String{Id: input.Id, Value: evaluated_string} // TODO evaluate here !!!!! get helper
-	} else if input.Default != nil {
-		cwl_obj = input.Default
-	} else {
-		err = fmt.Errorf("no input (source, default or valueFrom) defined for %s", input.Id)
-	}
-	obj = &cwl_obj
-	return
-}
-
-type WorkflowStepOutput struct {
-	Id string `yaml:"id"`
-}
-
-//http://www.commonwl.org/v1.0/Workflow.html#CommandLineBinding
-type CommandLineBinding struct {
-	LoadContents  bool   `yaml:"loadContents"`
-	Position      int    `yaml:"position"`
-	Prefix        string `yaml:"prefix"`
-	Separate      string `yaml:"separate"`
-	ItemSeparator string `yaml:"itemSeparator"`
-	ValueFrom     string `yaml:"valueFrom"`
-	ShellQuote    bool   `yaml:"shellQuote"`
-}
-
-// CreateWorkflowStepsArray
-func CreateWorkflowStepsArray(original interface{}) (err error, new_array []WorkflowStep) {
-
-
-	switch original.(type)  HERE
-
-	// iterate over workflow steps
-	for k, v := range original.(map[interface{}]interface{}) {
-		fmt.Printf("A step\n")
-		spew.Dump(v)
-
-		fmt.Println("type: ")
-		fmt.Println(reflect.TypeOf(v))
-
-		v_map := v.(map[interface{}]interface{})
-		spew.Dump(v_map)
-
-		switch v_map["in"].(type) {
-		case map[interface{}]interface{}:
-			err, v_map["in"] = CreateWorkflowStepInputArray(v_map["in"])
-			if err != nil {
-				return
-			}
-		}
-
-		switch v_map["out"].(type) {
-		case map[interface{}]interface{}: // example {'x'->?} -> [{id:'x'}]
-			fmt.Printf("match map[interface{}]interface{}\n")
-			err, v_map["out"] = CreateWorkflowStepOutputArray(v_map["out"])
-			if err != nil {
-				return
-			}
-		case []interface{}: // example ['x'] -> [{id:'x'}]
-			fmt.Printf("match []string\n")
-			err, v_map["out"] = CreateWorkflowStepOutputArray(v_map["out"])
-			if err != nil {
-				return
-			}
-		default:
-			// TODO some error
-			fmt.Printf("match default\n")
-			fmt.Println("type: ")
-			fmt.Println(reflect.TypeOf(v_map["out"]))
-		}
-
-		switch v_map["hints"].(type) {
-		case map[interface{}]interface{}:
-			// Convert map of outputs into array of outputs
-			v_map["hints"], err = CreateRequirementArray(v_map["hints"])
-			if err != nil {
-				return
-			}
-		}
-
-		switch v_map["requirements"].(type) {
-		case map[interface{}]interface{}:
-			// Convert map of outputs into array of outputs
-			v_map["requirements"], err = CreateRequirementArray(v_map["requirements"])
-			if err != nil {
-				return
-			}
-		}
-
-		var step WorkflowStep
-		err = mapstructure.Decode(v, &step)
-		if err != nil {
-			err = fmt.Errorf("(CreateWorkflowStepsArray) %s", err.Error())
-			return
-		}
-		step.Id = k.(string)
-
-		fmt.Printf("Last step\n")
-		spew.Dump(step)
-		//fmt.Printf("C")
-		new_array = append(new_array, step)
-		//fmt.Printf("D")
-
-	}
-	//spew.Dump(new_array)
-	return
-}
-
 func GetMapElement(m map[interface{}]interface{}, key string) (value interface{}, err error) {
 
 	for k, v := range m {
@@ -210,135 +40,6 @@ func GetMapElement(m map[interface{}]interface{}, key string) (value interface{}
 		}
 	}
 	err = fmt.Errorf("Element \"%s\" not found in map", key)
-	return
-}
-
-func CreateWorkflowStepInputArray(original interface{}) (err error, new_array []WorkflowStepInput) {
-
-	input_array, ok := original.(map[interface{}]interface{})
-	if !ok {
-		err = fmt.Errorf("could not parse WorkflowStepInputArray")
-		fmt.Println("original: ")
-		spew.Dump(original)
-		return
-	}
-
-	for k, v := range input_array {
-		//v is an input
-
-		var input_parameter WorkflowStepInput
-
-		fmt.Println("::::::::::::::::::::")
-		spew.Dump(v)
-		spew.Dump(input_parameter)
-
-		switch v.(type) {
-		case string:
-			fmt.Println("case string")
-			source_string := v.(string)
-			fmt.Printf("source_string: %s\n", source_string)
-			input_parameter.Source = []string{"#" + source_string} // TODO this is a  WorkflowStepInput.source or WorkflowStepInput (the latter would not make much sense)
-		case int:
-			fmt.Println("int")
-			input_parameter.Default = &Int{Id: input_parameter.Id, Value: v.(int)}
-		case map[interface{}]interface{}:
-			fmt.Println("case map[interface{}]interface{}")
-			err = mapstructure.Decode(v, &input_parameter)
-			if err != nil {
-				err = fmt.Errorf("(CreateWorkflowStepInputArray) %s", err.Error())
-				return
-			}
-
-			// set Default field
-			default_value, errx := GetMapElement(v.(map[interface{}]interface{}), "default")
-			if errx == nil {
-				switch default_value.(type) {
-				case string:
-					input_parameter.Default = &String{Id: input_parameter.Id, Value: default_value.(string)}
-				case int:
-					input_parameter.Default = &Int{Id: input_parameter.Id, Value: default_value.(int)}
-				default:
-					err = fmt.Errorf("string or int expected for key \"default\"")
-					return
-				}
-			}
-
-			// set ValueFrom field
-			valueFrom_if, errx := GetMapElement(v.(map[interface{}]interface{}), "valueFrom")
-			if errx == nil {
-				valueFrom_str, ok := valueFrom_if.(string)
-				if !ok {
-					err = fmt.Errorf("cannot convert valueFrom")
-					return
-				}
-				input_parameter.ValueFrom = Expression(valueFrom_str)
-			}
-
-		default:
-			fmt.Println("case default")
-			err = fmt.Errorf("Input type for %s can not be parsed", input_parameter.Id)
-			return
-		}
-
-		if input_parameter.Id == "" {
-			input_parameter.Id = k.(string)
-		}
-		fmt.Println("input_parameter:")
-		spew.Dump(input_parameter)
-
-		// now v should be a map
-
-		fmt.Println("done.")
-		new_array = append(new_array, input_parameter)
-		//fmt.Printf("D")
-
-	}
-	//spew.Dump(new_array)
-	return
-}
-
-func CreateWorkflowStepOutputArray(original interface{}) (err error, new_array []WorkflowStepOutput) {
-
-	switch original.(type) {
-	case map[interface{}]interface{}:
-
-		for k, v := range original.(map[interface{}]interface{}) {
-			//fmt.Printf("A")
-
-			var output_parameter WorkflowStepOutput
-			err = mapstructure.Decode(v, &output_parameter)
-			if err != nil {
-				err = fmt.Errorf("(CreateWorkflowStepOutputArray) %s", err.Error())
-				return
-			}
-
-			output_parameter.Id = k.(string)
-			//fmt.Printf("C")
-			new_array = append(new_array, output_parameter)
-			//fmt.Printf("D")
-
-		}
-
-	case []interface{}:
-		for _, v := range original.([]interface{}) {
-
-			switch v.(type) {
-			case string:
-				output_parameter := WorkflowStepOutput{Id: v.(string)}
-				new_array = append(new_array, output_parameter)
-			default:
-				wso, ok := v.(WorkflowStepOutput)
-				if !ok {
-					// TODO some ERROR
-				}
-				new_array = append(new_array, wso)
-			}
-
-		}
-
-	} // end switch
-
-	//spew.Dump(new_array)
 	return
 }
